## Changes

Some small changes to fw_devlinks supplier handling to include some dependencies it misses, and a faster method to find out the module name for a sysfs node:
- Add sysfs node parents' modules as dependencies (and parents of suppliers and so on)
- Do not limit supplier handling to the `platform` bus 
- Install modules for all suppliers of a supplier's module
- Use `driver/module` sysfs dirs for module name

Ends up cutting start-up time by half (~500ms) on ARM64. But it adds some start-up time (<100ms) on AMD64 because of devices outside `/sys/devices/platform` which this now handles.

<details><summary>Benchmarks on ARM64  (mt8183-kukui-jacuzzi-cozmo)</summary>

Before:
```
$ hyperfine -L count "250,125,50,10,1" './dracut-install -D "$(mktemp -d)" -o -m $(lsmod | cut -d" " -f1 | head -n {count})'
Benchmark 1: ./dracut-install -D "$(mktemp -d)" -o -m $(lsmod | cut -d" " -f1 | head -n 250)
  Time (mean ± σ):      3.710 s ±  0.030 s    [User: 2.299 s, System: 1.365 s]
  Range (min … max):    3.644 s …  3.750 s    10 runs
 
Benchmark 2: ./dracut-install -D "$(mktemp -d)" -o -m $(lsmod | cut -d" " -f1 | head -n 125)
  Time (mean ± σ):      2.947 s ±  0.015 s    [User: 1.930 s, System: 0.994 s]
  Range (min … max):    2.912 s …  2.964 s    10 runs
 
Benchmark 3: ./dracut-install -D "$(mktemp -d)" -o -m $(lsmod | cut -d" " -f1 | head -n 50)
  Time (mean ± σ):      2.394 s ±  0.026 s    [User: 1.621 s, System: 0.742 s]
  Range (min … max):    2.354 s …  2.437 s    10 runs
 
Benchmark 4: ./dracut-install -D "$(mktemp -d)" -o -m $(lsmod | cut -d" " -f1 | head -n 10)
  Time (mean ± σ):      1.255 s ±  0.009 s    [User: 1.004 s, System: 0.249 s]
  Range (min … max):    1.245 s …  1.274 s    10 runs
 
Benchmark 5: ./dracut-install -D "$(mktemp -d)" -o -m $(lsmod | cut -d" " -f1 | head -n 1)
  Time (mean ± σ):      1.021 s ±  0.007 s    [User: 0.888 s, System: 0.136 s]
  Range (min … max):    1.011 s …  1.030 s    10 runs
```

After:
```
$ hyperfine -L count "250,125,50,10,1" './dracut-install -D "$(mktemp -d)" -o -m $(lsmod | cut -d" " -f1 | head -n {count})'
Benchmark 1: ./dracut-install -D "$(mktemp -d)" -o -m $(lsmod | cut -d" " -f1 | head -n 250)
  Time (mean ± σ):      3.023 s ±  0.022 s    [User: 1.519 s, System: 1.450 s]
  Range (min … max):    2.994 s …  3.059 s    10 runs
 
Benchmark 2: ./dracut-install -D "$(mktemp -d)" -o -m $(lsmod | cut -d" " -f1 | head -n 125)
  Time (mean ± σ):      2.504 s ±  0.025 s    [User: 1.248 s, System: 1.214 s]
  Range (min … max):    2.461 s …  2.545 s    10 runs
 
Benchmark 3: ./dracut-install -D "$(mktemp -d)" -o -m $(lsmod | cut -d" " -f1 | head -n 50)
  Time (mean ± σ):      1.766 s ±  0.035 s    [User: 0.902 s, System: 0.840 s]
  Range (min … max):    1.719 s …  1.824 s    10 runs
 
Benchmark 4: ./dracut-install -D "$(mktemp -d)" -o -m $(lsmod | cut -d" " -f1 | head -n 10)
  Time (mean ± σ):     674.9 ms ±   8.9 ms    [User: 363.1 ms, System: 309.5 ms]
  Range (min … max):   665.8 ms … 698.9 ms    10 runs
 
Benchmark 5: ./dracut-install -D "$(mktemp -d)" -o -m $(lsmod | cut -d" " -f1 | head -n 1)
  Time (mean ± σ):     435.7 ms ±   3.1 ms    [User: 236.3 ms, System: 203.0 ms]
  Range (min … max):   430.5 ms … 438.5 ms    10 runs
```

</details>
<details><summary>Benchmarks on AMD64</summary>

Before:
```
$ hyperfine -L count "250,125,50,10,1" './dracut-install -D "$(mktemp -d)" -o -m $(lsmod | cut -d" " -f1 | head -n {count})'
Benchmark 1: ./dracut-install -D "$(mktemp -d)" -o -m $(lsmod | cut -d" " -f1 | head -n 250)
  Time (mean ± σ):      1.635 s ±  0.022 s    [User: 1.182 s, System: 0.459 s]
  Range (min … max):    1.610 s …  1.674 s    10 runs
 
Benchmark 2: ./dracut-install -D "$(mktemp -d)" -o -m $(lsmod | cut -d" " -f1 | head -n 125)
  Time (mean ± σ):     537.1 ms ±   4.8 ms    [User: 441.7 ms, System: 98.1 ms]
  Range (min … max):   529.1 ms … 544.5 ms    10 runs
 
Benchmark 3: ./dracut-install -D "$(mktemp -d)" -o -m $(lsmod | cut -d" " -f1 | head -n 50)
  Time (mean ± σ):     250.0 ms ±   1.8 ms    [User: 197.2 ms, System: 54.9 ms]
  Range (min … max):   245.8 ms … 252.8 ms    12 runs
 
Benchmark 4: ./dracut-install -D "$(mktemp -d)" -o -m $(lsmod | cut -d" " -f1 | head -n 10)
  Time (mean ± σ):      71.8 ms ±   0.9 ms    [User: 53.9 ms, System: 19.4 ms]
  Range (min … max):    69.1 ms …  75.3 ms    41 runs
 
Benchmark 5: ./dracut-install -D "$(mktemp -d)" -o -m $(lsmod | cut -d" " -f1 | head -n 1)
  Time (mean ± σ):      10.7 ms ±   0.3 ms    [User: 3.9 ms, System: 8.1 ms]
  Range (min … max):    10.2 ms …  12.4 ms    266 runs
```

After:
```
$ hyperfine -L count "250,125,50,10,1" './dracut-install -D "$(mktemp -d)" -o -m $(lsmod | cut -d" " -f1 | head -n {count})'
Benchmark 1: ./dracut-install -D "$(mktemp -d)" -o -m $(lsmod | cut -d" " -f1 | head -n 250)
  Time (mean ± σ):      1.836 s ±  0.039 s    [User: 1.211 s, System: 0.631 s]
  Range (min … max):    1.778 s …  1.884 s    10 runs
 
Benchmark 2: ./dracut-install -D "$(mktemp -d)" -o -m $(lsmod | cut -d" " -f1 | head -n 125)
  Time (mean ± σ):      1.670 s ±  0.029 s    [User: 1.118 s, System: 0.557 s]
  Range (min … max):    1.634 s …  1.716 s    10 runs
 
Benchmark 3: ./dracut-install -D "$(mktemp -d)" -o -m $(lsmod | cut -d" " -f1 | head -n 50)
  Time (mean ± σ):     323.8 ms ±   4.2 ms    [User: 217.9 ms, System: 108.0 ms]
  Range (min … max):   319.0 ms … 331.0 ms    10 runs
 
Benchmark 4: ./dracut-install -D "$(mktemp -d)" -o -m $(lsmod | cut -d" " -f1 | head -n 10)
  Time (mean ± σ):     143.5 ms ±   3.0 ms    [User: 73.0 ms, System: 72.0 ms]
  Range (min … max):   138.4 ms … 151.8 ms    20 runs
 
Benchmark 5: ./dracut-install -D "$(mktemp -d)" -o -m $(lsmod | cut -d" " -f1 | head -n 1)
  Time (mean ± σ):      71.6 ms ±   1.1 ms    [User: 13.6 ms, System: 59.4 ms]
  Range (min … max):    69.3 ms …  73.9 ms    40 runs
```
(The 1s jump at 125 modules is due to `amdgpu` now being pulled in as a supplier of `snd_hda_intel`)

</details>

## Checklist
- [x] I have tested it locally
- [ ] I have reviewed and updated any documentation if relevant
- [ ] I am providing new code and test(s) for it

Fixes #316 to a hopefully acceptable point, and fixes some issues I discovered while working on that.