@bdrung 2024-06-03
I tried this change on my desktop machine with an AMD Ryzen 7 5700G in a tmpfs chroot. Stock Ubuntu 24.04 (noble):

```
$ schroot-wrapper -p initramfs-tools,initramfs-tools-core,linux-image-generic,zstd,busybox-initramfs,cryptsetup-initramfs,kbd,lvm2,mdadm,ntfs-3g,plymouth,plymouth-theme-spinner,hyperfine -u root -c noble
(noble)root@desktop:~# hyperfine "update-initramfs -u"
Benchmark 1: update-initramfs -u
  Time (mean ± σ):      4.437 s ±  0.016 s    [User: 1.717 s, System: 3.155 s]
  Range (min … max):    4.417 s …  4.464 s    10 runs
```

I took dracut-install from Ubuntu 24.10 (oracular) and applied https://github.com/dracut-ng/dracut-ng/pull/328:

```
$ schroot-wrapper -p ../dracut-install_060+5-8ubuntu2_amd64.deb -p initramfs-tools,initramfs-tools-core,linux-image-generic,zstd,busybox-initramfs,cryptsetup-initramfs,kbd,lvm2,mdadm,ntfs-3g,plymouth,plymouth-theme-spinner,hyperfine -u root -c noble
(noble)root@desktop:~# hyperfine "update-initramfs -u"
Benchmark 1: update-initramfs -u
  Time (mean ± σ):      4.271 s ±  0.014 s    [User: 1.665 s, System: 3.047 s]
  Range (min … max):    4.254 s …  4.295 s    10 runs
```

Then added the patches from this merge request:

```
$ schroot-wrapper -p ../dracut-install_060+5-8ubuntu2_amd64.deb -p initramfs-tools,initramfs-tools-core,linux-image-generic,zstd,busybox-initramfs,cryptsetup-initramfs,kbd,lvm2,mdadm,ntfs-3g,plymouth,plymouth-theme-spinner,hyperfine -u root -c noble
(noble)root@desktop:~# hyperfine "update-initramfs -u"
Benchmark 1: update-initramfs -u
  Time (mean ± σ):      8.003 s ±  0.020 s    [User: 2.109 s, System: 6.330 s]
  Range (min … max):    7.983 s …  8.045 s    10 runs
```

The execution time doubled! Even after reducing the number of dracut-install calls in initramfs-tools, it is still significant slower:

```
$ schroot-wrapper  -p ../dracut-install_060+5-8ubuntu2_amd64.deb -p ../initramfs-tools_0.142ubuntu27_all.deb,../initramfs-tools-core_0.142ubuntu27_all.deb,linux-image-generic,zstd,busybox-initramfs,cryptsetup-initramfs,kbd,lvm2,mdadm,ntfs-3g,plymouth,plymouth-theme-spinner,hyperfine -u root -c oracular
(oracular)root@desktop:~# hyperfine "update-initramfs -u"
Benchmark 1: update-initramfs -u
  Time (mean ± σ):      5.079 s ±  0.011 s    [User: 1.764 s, System: 3.764 s]
  Range (min … max):    5.061 s …  5.091 s    10 runs
```